+++ docs/HISTORY.rst
@@ -6,7 +6,7 @@ Change Log
 
 **Features**
 
-- Added a ``timeout`` parameter to ``listen_state()`` to delete the callback after a pre-determined interval.
+- Added a ``timeout`` parameter to ``listen_state()`` and ``listen_event()`` to delete the callback after a pre-determined interval.
 
 **Fixes**
 **Breaking Changes**